fix(logs): don't create log file on web

+++ lib/logs/logging_manager.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:flutter/foundation.dart';
 import 'package:helpers/helpers.dart';
 
 /// `FileOutput` import needed due to bug in package.
@@ -16,7 +17,7 @@ late final Logger log;
 /// Manages logging for the app.
 class LoggingManager {
   /// The file to which logs are saved.
-  final File _logFile;
+  final File? _logFile;
 
   /// Singleton instance for easy access.
   static late final LoggingManager instance;
@@ -34,15 +35,19 @@ class LoggingManager {
       return LoggingManager._(File(''));
     }
 
-    final dataDir = await getApplicationSupportDirectory();
-    final logFile = File('${dataDir.path}${Platform.pathSeparator}log.txt');
-    if (await logFile.exists()) await logFile.delete();
-    await logFile.create();
-
     final List<LogOutput> outputs = [
       ConsoleOutput(),
-      FileOutput(file: logFile),
     ];
+
+    // Create a log file for the current run.
+    File? logFile;
+    if (!kIsWeb) {
+      final dataDir = await getApplicationSupportDirectory();
+      logFile = File('${dataDir.path}${Platform.pathSeparator}log.txt');
+      if (await logFile.exists()) await logFile.delete();
+      await logFile.create();
+      outputs.add(FileOutput(file: logFile));
+    }
 
     log = Logger(
       filter: ProductionFilter(),
@@ -60,7 +65,7 @@ class LoggingManager {
   }
 
   /// Read the logs from this run from the log file.
-  Future<String> getLogs() async => await _logFile.readAsString();
+  Future<String?> getLogs() async => await _logFile?.readAsString();
 
   /// Close the logger and release resources.
   void close() => log.close();
